refactor: 불필요한 TabProvider 제거 #18
